Add files via upload
This script was used in the following study:

Ondrej Belan et al., "Single-molecule analysis reveals cooperative stimulation of Rad51 filament nucleation and growth by mediator proteins", Molecular Cell, 2021, https://doi.org/10.1016/j.molcel.2020.12.020.

In this script, protein filaments that nucleate and elongate on a tethered DNA are tracked on a kymograph.

The image (imported as a raw .tiff file) is cropped, median-filtered, and smoothed. Subsequently, a peak-finder function scans each timeframe (vertical line) of the kymograph. The number of detected peaks is then plotted as a function of time. Finally, the dependency is fitted with a single exponential function to obtain a nucleation rate.